# Perf/less rx operators